refactor: roomList DIR 구조 변경 #25
- rooms로 통일
