+++ app/rooms/_components/RoomEmpty.tsx
