+++ app/rooms/_components/RoomItem.tsx
@@ -11,8 +11,8 @@ import {
 } from "react-native";
 import { Card, Text, useTheme } from "react-native-paper";
 import { RFValue } from "react-native-responsive-fontsize";
-import { backgroundColor, color } from "../../../common/styles/color";
-import { Room } from "../types/dto/responses/room";
+import { backgroundColor, color } from "../../../src/common/styles/color";
+import { Room } from "../../../src/domain/rooms/types/room";
 import RoomInfoSheet from "./sheets/RoomOptionSheet";
 
 interface RoomItemProps {
