+++ package.json
@@ -7,7 +7,8 @@
     "android": "expo start --android",
     "ios": "expo start --ios",
     "web": "expo start --web",
-    "mock": "node ../mock-server-api/mocks/server.js"
+    "mock": "node ../mock-server-api/mocks/server.js",
+    "build-preview": "eas build --platform android --profile preview"
   },
   "dependencies": {
     "@react-navigation/bottom-tabs": "^7.2.0",
