+++ src/domain/rooms/hooks/mutations/roomMutations.ts
@@ -1,7 +1,7 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
-import roomService from "../../services/roomListService";
+import roomService from "../../services/roomService";
 import QUERY_KEYS from "../../../../common/hooks/queries/queryKeys";
-import { Room } from "../../types/dto/responses/room";
+import { Room } from "../../types/room";
 
 // 방 알림 설정 토글 mutation
 export const useToggleRoomNotificationMutation = () => {
