+++ src/domain/rooms/types/room.ts
